+ added SimplePagerFragment + added onSkipButtonClicked method
