+++ lib/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">SlidingTutorialLibrary</string>
+    <string name="skip">Skip</string>
 </resources>
